Refactor ecdict connector
