+++ src/cmdict/pdf_tools.py
@@ -1,7 +1,7 @@
 """Functions to handle highlights in PDF files."""
 import fitz
 
-from cmdict.db_connector import DBConnector
+from cmdict.ecdict_connector import ecdict_engine
 from cmdict.utils import remove_punctuation
 
 
@@ -151,7 +151,7 @@ def _fix_hyphen_broken(word_list):
         w1, w2 = word_list[i], word_list[i + 1]
         combined = w1[:-1] + w2
 
-        if w1.endswith("-") and DBConnector().query(combined):
+        if w1.endswith("-") and ecdict_engine.query(combined):
             res.append(combined)
             i += 1
         else:
